Reorganize calibrate methods
Moved calibrate_ methods into crystal_calibrate
